% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gower.R
\name{discrete.dist}
\alias{discrete.dist}
\title{Apply the discrete metric to two scores.}
\usage{
discrete.dist(x, y)
}
\arguments{
\item{x}{a score.}

\item{y}{a score.}
}
\value{
0 if \code{x} is equal to \code{y}, 1 if \code{x} is not equal to \code{y}, \code{NA} if either score is \code{NA}.
}
\description{
Apply the discrete metric to two scores.
}
\details{
This function applies the discrete metric to two scores. This is an appropriate distance function for nominal data.
}
\seealso{
\code{\link{L1.dist}}
}
