% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gower.R
\name{hpd}
\alias{hpd}
\title{Compute a highest posterior density (HPD) interval.}
\usage{
hpd(x, alpha = 0.05)
}
\arguments{
\item{x}{the posterior sample.}

\item{alpha}{the desired significance level.}
}
\value{
A 2-vector containing the lower endpoint and the upper endpoint, respectively.
}
\description{
Compute a highest posterior density (HPD) interval.
}
\details{
This function uses a given posterior sample to compute an HPD interval at a given significance level.
}
