% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gower.R
\name{L1.dist}
\alias{L1.dist}
\title{Apply the L1 distance function to two scores.}
\usage{
L1.dist(x, y, range)
}
\arguments{
\item{x}{a score.}

\item{y}{a score.}

\item{range}{the range of the scores, i.e., the difference between the maximum score and the minimum score.}
}
\value{
\eqn{|x - y| / r}, where \eqn{r} is the range of the scores; or \code{NA} if either score is \code{NA}.
}
\description{
Apply the L1 distance function to two scores.
}
\details{
This function applies the L1 distance function to two scores. This is an appropriate distance function for ordinal data.
}
\seealso{
\code{\link{discrete.dist}}
}
