% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liver}
\alias{liver}
\title{Ordinal data from a radiological study of congenital diaphragmatic hernia.}
\format{
A data frame having 47 rows and four columns
}
\usage{
data(liver)
}
\description{
This data frame has exactly four columns. The first two columns contain liver herniation grades assigned by the first radiologist, and the final two columns contain grades assigned by the second radiologist. Each radiologist graded each image twice so that intra-rater consistency can be assessed. The liver grades and their meanings are provided below.
}
\details{
\itemize{
  \item 1. No herniation of the liver into the fetal chest
  \item 2. Less than half of the ipsilateral thorax is occupied by the fetal liver
  \item 3. Greater than half of the thorax is occupied by the fetal liver
  \item 4. The liver dome reaches the thoracic apex
  \item 5. The liver dome not only reaches the thoracic apex but also extends across the thoracic midline
}
}
\keyword{datasets}
